+++ chia/wallet/wallet_state_manager.py
@@ -890,14 +890,14 @@ class WalletStateManager:
         return [tx_record.name]
 
     async def filter_spam(self, new_coin_state: List[CoinState]) -> List[CoinState]:
-        kop_spam_amount = self.config.get("kop_spam_amount", 1000000)
+        cch_spam_amount = self.config.get("cch_spam_amount", 1000000)
 
         # No need to filter anything if the filter is set to 1 or 0 mojos
-        if kop_spam_amount <= 1:
+        if cch_spam_amount <= 1:
             return new_coin_state
 
         spam_filter_after_n_txs = self.config.get("spam_filter_after_n_txs", 200)
-        small_unspent_count = await self.coin_store.count_small_unspent(kop_spam_amount)
+        small_unspent_count = await self.coin_store.count_small_unspent(cch_spam_amount)
 
         # if small_unspent_count > spam_filter_after_n_txs:
         filtered_cs: List[CoinState] = []
@@ -908,7 +908,7 @@ class WalletStateManager:
             if (
                 cs.created_height is not None
                 and cs.spent_height is None
-                and cs.coin.amount < kop_spam_amount
+                and cs.coin.amount < cch_spam_amount
                 and (cs.coin.puzzle_hash in is_standard_wallet_phs or await self.is_standard_wallet_tx(cs))
             ):
                 is_standard_wallet_phs.add(cs.coin.puzzle_hash)